fix(config): renames __init__ imports
