+++ archai/discrete_search/search_spaces/config/__init__.py
@@ -1,5 +1,5 @@
-from archai.discrete_search.search_spaces.builder.discrete_choice import DiscreteChoice
-from archai.discrete_search.search_spaces.builder.arch_param_tree import ArchParamTree
-from archai.discrete_search.search_spaces.builder.arch_config import ArchConfig, ArchConfigList, build_arch_config
-from archai.discrete_search.search_spaces.builder.search_space import SearchSpaceBuilder
-from archai.discrete_search.search_spaces.builder.helpers import repeat_config
+from archai.discrete_search.search_spaces.config.discrete_choice import DiscreteChoice
+from archai.discrete_search.search_spaces.config.arch_param_tree import ArchParamTree
+from archai.discrete_search.search_spaces.config.arch_config import ArchConfig, ArchConfigList, build_arch_config
+from archai.discrete_search.search_spaces.config.search_space import ConfigSearchSpace
+from archai.discrete_search.search_spaces.config.helpers import repeat_config
